+++ database/migrations/2025_09_04_234743_create_datos_salidas_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('datos_salidas', function (Blueprint $table) {
             $table->id();
             $table->dateTime('fecha');
-            $table->decimal('kilometraje',10,2);
+            $table->unsignedInteger('kilometraje');
             $table->foreignId('gasolina')->constrained('niveles_combustible');
             $table->foreignId('orden_servicio_id')->constrained('ordenes_servicio');
         });
